pyproject.toml
[tool.poetry]
name = "summery_creator"
version = "0.1.0"
description = ""
authors = ["Alexandr Karpov <alexandr.d.karpov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.5"
pytz = "^2022.7"
psutil = "^5.9.4"
python-slugify = "^7.0.0"
pillow = "^9.3.0"
argon2-cffi = "^21.3.0"
whitenoise = "^6.3.0"
redis = "^4.4.0"
celery = {extras = ["redis"], version = "^5.2.7"}
django-celery-beat = "^2.4.0"
flower = "^1.2.0"
django = "^4.1.5"
django-structlog = "^4.0.1"
django-environ = "^0.9.0"
django-model-utils = "^4.3.1"
django-redis = "^5.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.25.1"
werkzeug = {extras = ["watchdog"], version = "^2.2.2"}
ipdb = "^0.13.11"
watchfiles = "^0.18.1"
mypy = "^0.991"
django-stubs = "^1.13.1"
pytest = "^7.2.0"
pytest-sugar = "^0.9.6"
djangorestframework-stubs = "^1.8.0"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"
coverage = "^7.0.3"
black = "^22.12.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
pre-commit = "^2.21.0"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
django-coverage-plugin = "^3.0.0"
pytest-django = "^4.5.2"
sentry-sdk = "^1.12.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
requests-toolbelt = "^1.0.0"
python-magic = "^0.4.27"
moviepy = "^1.0.3"
textract = "^1.6.5"
django-elasticsearch-dsl = "^8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
